separate user and document domains
