+++ src/models/index.js
@@ -1,0 +1,7 @@
+import Users from './user';
+import Documents from './document';
+
+export default {
+  Users,
+  Documents
+};
