+++ src/resolvers/index.js
@@ -1,0 +1,4 @@
+import userResolvers from './user';
+import documentResolvers from './document';
+
+export default [userResolvers, documentResolvers];
