+++ src/resolvers/user.js
@@ -1,9 +1,8 @@
 import { UserInputError, AuthenticationError } from 'apollo-server-core';
-import Users from './user.schema';
-import Documents from './document.schema';
+import Users from '../models/user';
 import { createToken } from '../util';
 
-const resolvers = {
+export default {
   Query: {
     getUser: async (_, args) => {
       try {
@@ -13,31 +12,18 @@ const resolvers = {
         return new Error(error.message);
       }
     },
-    getDocument: async (_, args) => {
-      try {
-        const response = await Documents.findById(args.id);
-        return response;
-      } catch (error) {
-        return new Error(error.message);
-      }
-    },
     getUsers: async (_, args, context) => {
-      if (context.user.role !== 'ADMIN') {
-        throw new Error(
-          'Permission denied. Only Admins can query for all users'
-        );
-      }
+      // if (context.user.role !== 'ADMIN') {
+      //   throw new Error(
+      //     'Permission denied. Only Admins can query for all users'
+      //   );
+      // }
       return await Users.find({})
         .populate('documents')
         .exec();
-    },
-    getDocuments: async () => {
-      const docs = await Documents.find({})
-        .populate('owner')
-        .exec();
-      return docs;
     }
   },
+
   Mutation: {
     registerUser: async (_, args) => {
       try {
@@ -102,48 +88,10 @@ const resolvers = {
       } catch (error) {
         throw new Error(error.message);
       }
-    },
-    createDocument: async (parent, args, context) => {
-      try {
-        const { title, content, access = 'PUBLIC' } = args;
-        const newDocument = await Documents.create({
-          title,
-          content,
-          access,
-          owner: context.user.id
-        });
-        return newDocument;
-      } catch (error) {
-        throw new Error(error);
-      }
-    },
-    updateDocument: async (_, args) => {
-      try {
-        const updatedDocument = await Documents.findByIdAndUpdate(
-          args.id,
-          args,
-          { new: true }
-        );
-        if (!updatedDocument) {
-          throw new Error('Document could not be updated');
-        }
-        return updatedDocument;
-      } catch (error) {
-        throw new Error(error.message);
-      }
-    },
-    deleteDocument: async (_, args) => {
-      try {
-        const documentFound = await Documents.findByIdAndDelete(args.id);
-        if (documentFound) {
-          return { message: 'Document deleted' };
-        }
-        throw new Error('Document does not exist');
-      } catch (error) {
-        throw new Error(error.message);
-      }
     }
+  },
+
+  User: {
+    documents: async (user, args, context) => {}
   }
 };
-
-export default resolvers;
