--- src/schema/index.js
@@ -1,2 +1,0 @@
-export { default as resolvers } from './resolvers';
-export { default as typeDefs } from './typeDefs';
